Milestone 1: Getting started
Load the router library.
Add a nav bar to the shell template with anchor tags, routerLink and routerLinkActive directives.
Add a router-outlet to the shell template where views will be displayed.
Configure the router module with RouterModule.forRoot.
Set the router to compose HTML5 browser URLs.
handle invalid routes with a wildcard route.
navigate to the default route when the app launches with an empty path.